Eliminate isspace() related compilation warnings
git-svn-id: http://casvn/svn/octeon/bdk/branches/thunder.ddr4_rdimm@115462 c83883e8-9f28-0410-baf1-f2025811e16a

+++ libdram/dram-spd.c
@@ -1,4 +1,5 @@
 #include <bdk.h>
+#include <ctype.h>
 #include "dram-internal.h"
 
 
@@ -306,7 +307,7 @@ int get_dimm_part_number(char *buffer, bdk_node_t node, const dimm_config_t *dim
     --strlen; /* point to the last character */
 
     /* Remove trailing spaces */
-    while (isspace(buffer[strlen]) && strlen) {
+    while (isspace((int)buffer[strlen]) && strlen) {
         --strlen;
     }
